scrapy_app/requirements.txt
scrapy==2.11.1
psycopg2-binary